+++ module/schemas/commonSchemas.js
@@ -49,6 +49,9 @@ export const itemBase = () => {
   };
 };
 
+export const basicTextField = () =>
+  new fields.StringField({ required: false, blank: true, initial: "" });
+
 export const possibleRanges = Object.keys(ARM5E.magic.ranges).filter(r => !r.disabled);
 export const possibleTargets = Object.keys(ARM5E.magic.targets).filter(r => !r.disabled);
 export const possibleDurations = Object.keys(ARM5E.magic.durations).filter(r => !r.disabled);